Fix JPA id type mismatch with SQLite integer columns

+++ toluja-app/backend/src/main/java/com/toluja/app/dto/OrderDtos.java
@@ -10,11 +10,11 @@ import java.util.List;
 
 public class OrderDtos {
     public record CreateOrderRequest(String observacao, List<@Valid CreateOrderItemRequest> itens) {}
-    public record CreateOrderItemRequest(@NotNull Long itemId, @NotNull @Min(1) Integer quantidade) {}
+    public record CreateOrderItemRequest(@NotNull Integer itemId, @NotNull @Min(1) Integer quantidade) {}
 
-    public record OrderResponse(Long id, String codigo, OffsetDateTime criadoEm, String status, BigDecimal total,
+    public record OrderResponse(Integer id, String codigo, OffsetDateTime criadoEm, String status, BigDecimal total,
                                 String observacao, UserSummary user, List<OrderItemResponse> itens) {}
-    public record UserSummary(Long id, @NotBlank String username, String nomeExibicao) {}
-    public record OrderItemResponse(Long id, Long itemId, String nomeSnapshot, BigDecimal precoSnapshot,
+    public record UserSummary(Integer id, @NotBlank String username, String nomeExibicao) {}
+    public record OrderItemResponse(Integer id, Integer itemId, String nomeSnapshot, BigDecimal precoSnapshot,
                                     Integer quantidade, BigDecimal subtotal) {}
 }
